app/requirements.txt
PyGithub
nltk
flask
requests
flask-cors
Click==7.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2>=2.10.1
MarkupSafe==1.1.0
Werkzeug==0.14.1